requirements.txt
Flask==2.0.3
Werkzeug<2.1
Pillow
gTTS
paddleocr
paddlepaddle
